requirements.txt
streamlit==1.28.1
pandas==2.1.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0
reportlab==4.0.7
python-dotenv==1.0.0
bcrypt==4.0.1
cryptography==38.0.4
